Removed useless resolvers/shared/util.ts
