+++ src/schema/resolvers/shared/account.ts
@@ -1,6 +1,6 @@
 import { db } from "../../../database";
 import { Account, AccountID } from "../../../model";
-import { createBatchResolver, idOnlyRequested } from "./util";
+import { createBatchResolver, idOnlyRequested } from "../util";
 
 export const account = createBatchResolver<any, Account[]>((source: any, args: any, context: any, info: any) => {
   const ids: AccountID[] = source.map((s: any) => s[info.fieldName]);
